requirements-dev.txt
autopep8==1.3.4
flake8==3.5.0
nose==1.3.7
singer-tools==0.4.1
requests-mock==1.4.0
